+++ src/test/java/TestNGFramework/FirstTest.java
@@ -22,7 +22,7 @@ public  class FirstTest {
        
     }	
 	
-	@Test(invocationCount = 10)
+	@Test(invocationCount=100)
 	public void FirstTest2() throws InterruptedException {
 		
 	    driver = new ChromeDriver();
@@ -39,29 +39,29 @@ public  class FirstTest {
 		Thread.sleep(2000);
 		driver.findElement(By.id("attachCV")).sendKeys("C:\\Users\\Suresh Bathula\\Downloads\\Suresh-resume1.pdf");
 		Thread.sleep(4000);
-		driver.quit();
-		FirstTest1();
+		//driver.quit();
+		Thread.sleep(20000);
 	   }
-	@Test
-public  void FirstTest1() throws InterruptedException {
-		
-		driver = new ChromeDriver();
-		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
-		driver.get("https://www.naukri.com/");
-		driver.manage().window().maximize();
-		driver.findElement(By.xpath("//*[text()='Login']")).click();
-		
-		driver.findElement(By.xpath("//*[@placeholder='Enter your active Email ID / Username']")).sendKeys("bathulasuresh518@gmail.com");
-		driver.findElement(By.xpath("//*[@placeholder='Enter your password']")).sendKeys("Suresh@1998");
-		driver.findElement(By.xpath("(//*[@type='submit'])[1]")).click();
-		//driver.findElement(By.xpath("//*[@class='crossIcon chatBot chatBot-ic-cross']")).click();
-		driver.findElement(By.xpath("//*[text()=' profile']")).click();
-		Thread.sleep(2000);
-		driver.findElement(By.id("attachCV")).sendKeys("C:\\Users\\Suresh Bathula\\Desktop\\Suresh_Automation_ResumeCV.pdf");
-		Thread.sleep(5000);
-		FirstTest2();
-		
-	}
+	
+//public  void FirstTest1() throws InterruptedException {
+//		
+//		driver = new ChromeDriver();
+//		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+//		driver.get("https://www.naukri.com/");
+//		driver.manage().window().maximize();
+//		driver.findElement(By.xpath("//*[text()='Login']")).click();
+//		
+//		driver.findElement(By.xpath("//*[@placeholder='Enter your active Email ID / Username']")).sendKeys("bathulasuresh518@gmail.com");
+//		driver.findElement(By.xpath("//*[@placeholder='Enter your password']")).sendKeys("Suresh@1998");
+//		driver.findElement(By.xpath("(//*[@type='submit'])[1]")).click();
+//		//driver.findElement(By.xpath("//*[@class='crossIcon chatBot chatBot-ic-cross']")).click();
+//		driver.findElement(By.xpath("//*[text()=' profile']")).click();
+//		Thread.sleep(2000);
+//		driver.findElement(By.id("attachCV")).sendKeys("C:\\Users\\Suresh Bathula\\Desktop\\Suresh_Automation_ResumeCV.pdf");
+//		Thread.sleep(5000);
+//		FirstTest2();
+//		
+//	}
 	@AfterMethod
     public void tearDown() {
         if (driver != null) {
