+++ test-output/NaukriSuite/NaukriTest.html
@@ -1,6 +1,6 @@
 <html>
 <head>
-<title>TestNG:  NaukriTests</title>
+<title>TestNG:  NaukriTest</title>
 <link href="../testng.css" rel="stylesheet" type="text/css" />
 <link href="../my-testng.css" rel="stylesheet" type="text/css" />
 
@@ -53,13 +53,13 @@ function toggleAllBoxes() {
 
 </head>
 <body>
-<h2 align='center'>NaukriTests</h2><table border='1' align="center">
+<h2 align='center'>NaukriTest</h2><table border='1' align="center">
 <tr>
 <td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
 </tr><tr>
 <td>Started on:</td><td>Thu Sep 18 11:16:24 IST 2025</td>
 </tr>
-<tr><td>Total time:</td><td>18 seconds (18437 ms)</td>
+<tr><td>Total time:</td><td>19 seconds (19788 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -76,10 +76,10 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='TestNGFramework.FirstTest.FirstTest2()'><b>FirstTest2</b><br>Test class: TestNGFramework.FirstTest</td>
+<td title='TestNGFramework.FirstTest2.FirstTest1()'><b>FirstTest1</b><br>Test class: TestNGFramework.FirstTest2</td>
 <td></td>
-<td>18</td>
-<td>TestNGFramework.FirstTest@7b98f307</td></tr>
+<td>19</td>
+<td>TestNGFramework.FirstTest2@40005471</td></tr>
 </table><p>
 </body>
 </html>
