+++ test-output/NaukriSuite/NaukriTest.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitXMLReporter -->
+<testsuite ignored="0" hostname="19FMUCHH010209" failures="0" tests="1" name="NaukriTest" time="19.788" errors="0" timestamp="2025-09-18T11:16:44 IST">
+  <testcase classname="TestNGFramework.FirstTest2" name="FirstTest1" time="19.459"/>
+</testsuite> <!-- NaukriTest -->
