+++ test-output/junitreports/TEST-TestNGFramework.FirstTest2.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="19FMUCHH010209" failures="0" tests="1" name="TestNGFramework.FirstTest2" time="19.756" errors="0" timestamp="2025-09-18T11:16:44 IST" skipped="0">
+  <testcase classname="TestNGFramework.FirstTest2" name="FirstTest1" time="19.756"/>
+</testsuite> <!-- TestNGFramework.FirstTest2 -->
